+++ src/Models/Camera.php
@@ -5,6 +5,8 @@ namespace TromsFylkestrafikk\Camera\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use TromsFylkestrafikk\Camera\Services\CameraTokenizer;
 
@@ -50,5 +52,42 @@ class Camera extends Model
     public function getFilePathRegexAttribute()
     {
         return preg_quote($this->folderPath) . '/' . $this->fileRegex;
+    }
+
+    /**
+     * Get the full path to the camera's current file.
+     *
+     * @return string
+     */
+    public function getCurrentPathAttribute()
+    {
+        return $this->currentFile ? $this->folderPath . '/' . $this->currentFile : null;
+    }
+
+    /**
+     * Get a decent cache key for this model
+     *
+     * @return string
+     */
+    public function getCacheKeyAttribute()
+    {
+        return "tromsfylkestrafikk.camera.{$this->id}";
+    }
+
+    public function getCurrentCacheKeyAttribute()
+    {
+        return $this->cacheKey . '.current';
+    }
+
+    /**
+     * Add cache around current image when set.
+     */
+    public function setCurrentFileAttribute($image)
+    {
+        $timeout = config('camera.cache_current');
+        if ($timeout) {
+            Cache::put($this->currentCacheKey, $image, config('camera.cache_current'));
+        }
+        $this->attributes['currentFile'] = $image;
     }
 }